+++ packages/backend/src/database/migrations/20240424131220_populate_slugs.ts
@@ -24,26 +24,26 @@ export async function up(knex: Knex): Promise<void> {
 
     // Make slugs notNullable
     await knex.schema.alterTable('saved_queries', (table) => {
-        table.string('slug').notNullable().alter();
+        table.text('slug').notNullable().alter();
     });
     await knex.schema.alterTable('dashboards', (table) => {
-        table.string('slug').notNullable().alter();
+        table.text('slug').notNullable().alter();
     });
     await knex.schema.alterTable('spaces', (table) => {
-        table.string('slug').notNullable().alter();
+        table.text('slug').notNullable().alter();
     });
 }
 
 export async function down(knex: Knex): Promise<void> {
     // Make slugs nullable
     await knex.schema.alterTable('saved_queries', (table) => {
-        table.string('slug').nullable().alter();
+        table.text('slug').nullable().alter();
     });
     await knex.schema.alterTable('dashboards', (table) => {
-        table.string('slug').nullable().alter();
+        table.text('slug').nullable().alter();
     });
     await knex.schema.alterTable('spaces', (table) => {
-        table.string('slug').nullable().alter();
+        table.text('slug').nullable().alter();
     });
 
     await knex.raw(`UPDATE saved_queries