+++ packages/backend/src/services/DashboardService/DashboardService.mock.ts
@@ -201,8 +201,9 @@ export const createDashboardWithSlug = {
     slug: 'my-new-dashboard',
 };
 
-export const createDashboardWithTileIds: CreateDashboard = {
+export const createDashboardWithTileIds: CreateDashboard & { slug: string } = {
     ...createDashboard,
+    slug: 'my-new-dashboard',
     tiles: [createTileWithId],
 };