+++ packages/common/src/types/featureFlags.ts
@@ -24,11 +24,6 @@ export enum FeatureFlags {
     /**/
     CustomSQLEnabled = 'custom-sql-enabled',
 
-    /**/
-    PuppeteerScrollElementIntoView = 'puppeteer-scroll-element-into-view',
-    PuppeteerSetViewportDynamically = 'puppeteer-set-viewport-dynamically',
-    PuppeteerDisconnectBrowser = 'puppeteer-disconnect-browser',
-
     /* Shows the two-stage login flow */
     newLoginEnabled = 'new-login-enabled',
 
@@ -49,4 +44,6 @@ export enum FeatureFlags {
 
     /* Send local timezone to the warehouse session */
     EnableUserTimezones = 'enable-user-timezones',
+
+    PromoteCharts = 'promote-charts',
 }