+++ packages/common/src/types/savedCharts.ts
@@ -620,3 +620,8 @@ export type ApiCalculateTotalResponse = {
     status: 'ok';
     results: Record<string, number>;
 };
+
+export type ApiPromoteChartResponse = {
+    status: 'ok';
+    results: SavedChartDAO;
+};