+++ packages/common/src/types/space.ts
@@ -29,6 +29,7 @@ export type SpaceSummary = Pick<
     | 'isPrivate'
     | 'pinnedListUuid'
     | 'pinnedListOrder'
+    | 'slug'
 > & {
     userAccess: SpaceShare | undefined;
     access: string[];