+++ packages/frontend/src/pages/ProjectSettings.tsx
@@ -6,6 +6,7 @@ import ErrorState from '../components/common/ErrorState';
 import PageBreadcrumbs from '../components/common/PageBreadcrumbs';
 import SuboptimalState from '../components/common/SuboptimalState/SuboptimalState';
 import CustomSqlPanel from '../components/CustomSqlPanel/CustomSqlPanel';
+import { DataOps } from '../components/DataOps';
 import DbtCloudSettings from '../components/DbtCloudSettings';
 import ProjectUserAccess from '../components/ProjectAccess';
 import { UpdateProjectConnection } from '../components/ProjectConnection';
@@ -109,6 +110,13 @@ const ProjectSettings: FC = () => {
                         path={`/generalSettings/projectManagement/${projectUuid}/customSql`}
                     >
                         <CustomSqlPanel projectUuid={projectUuid} />
+                    </Route>
+
+                    <Route
+                        exact
+                        path={`/generalSettings/projectManagement/${projectUuid}/dataOps`}
+                    >
+                        <DataOps projectUuid={projectUuid} />
                     </Route>
 
                     <Redirect to={`/generalSettings/`} />